mm: introduce sysfs interface for debugging kernel shrinker
This commit introduces the /sys/kernel/shrinker sysfs interface
which provides an ability to observe the state and interact with
individual kernel memory shrinkers.

Because the feature is oriented on kernel developers and adds some
memory overhead (which shouldn't be large unless there is a huge
amount of registered shrinkers), it's guarded by a config option
(disabled by default).

To simplify the code, kobjects are not embedded into shrinkers
objects, but are created, linked and unlinked dynamically.

This commit introduces basic "count" and "scan" interfaces.
Basic usage:
  $ cat count                   : get the number of objects
  $ echo "500" > scan           : try to reclaim 500 objects
  $ cat scan                    : get the number of objects reclaimed
Following commits in the series will add memcg- and numa-specific
features.

This commit gives sysfs entries simple numeric names, which are not
very convenient. The following commit in the series will provide
shrinkers with more meaningful names.

Signed-off-by: Roman Gushchin <roman.gushchin@linux.dev>