mm: kmem: add direct objcg pointer to task_struct
To charge a freshly allocated kernel object to a memory cgroup, the
kernel needs to obtain an objcg pointer. Currently it does it
indirectly by obtaining the memcg pointer first and then calling to
__get_obj_cgroup_from_memcg().

Usually tasks spend their entire life belonging to the same object
cgroup. So it makes sense to save the objcg pointer on task_struct
directly, so it can be obtained faster. It requires some work on fork,
exit and cgroup migrate paths, but these paths are way colder.

The old indirect way is still used for remote memcg charging.

This patch improves the performance of a kernel allocation
by additional 6-26% on top of the first patch in the series
as measured by a simple test listed below.

| version                | accounted (us) | delta | unaccounted (us) |  delta |
|------------------------+----------------+-------+------------------+--------|
| baseline (6.1+)        |          81042 |       |            45269 |        |
| optimize obj_cgroup... |          78756 | -2.8% |            42731 |  -5.6% |
| this patch             |          73650 | -9.1% |            30662 | -32.3% |

static int memory_alloc_test(struct seq_file *m, void *v)
{
        unsigned long i, j;
        void **ptrs;
        ktime_t start, end;
        s64 delta, min_delta = LLONG_MAX;

        ptrs = kvmalloc(sizeof(void *) * 1024 * 1024, GFP_KERNEL);
        if (!ptrs)
                return -ENOMEM;

        for (j = 0; j < 100; j++) {
                start = ktime_get();
                for (i = 0; i < 1024 * 1024; i++)
                        ptrs[i] = kmalloc(8, GFP_KERNEL_ACCOUNT);
                end = ktime_get();

                delta = ktime_us_delta(end, start);
                if (delta < min_delta)
                        min_delta = delta;

                for (i = 0; i < 1024 * 1024; i++)
                        kfree(ptrs[i]);
        }

        kvfree(ptrs);
        seq_printf(m, "%lld us\n", min_delta);

        return 0;
}

Unaccounted allocations were performed by a task belonging to the root
memory cgroup.

Signed-off-by: Roman Gushchin <roman.gushchin@linux.dev>